feat: RoiDynamic
