+++ src/components/ApyCalculator/CurrencyInputRoi/NumericalInput.tsx
