+++ src/components/ApyCalculator/DetailsContent.tsx
