--- src/components/RoiCalculator/index.tsx
@@ -1,1 +1,0 @@
-export { default as Roicalculator } from './RoiCalculator'
